drawing text's with node (label)
