+++ engine/public/core/ember_core.h
@@ -3,6 +3,7 @@
 #include "core/component/camera.h"
 #include "core/component/node.h"
 #include "core/component/sprite_node.h"
+#include "core/component/label.h"
 #include "core/engine_structs.h"
 
 #pragma region OPENGL/ES
@@ -26,6 +27,7 @@
  */
 class Renderer {
 public:
+    virtual ~Renderer() = default;
     Renderer() = default;
 
     SDL_Window* Window = nullptr;
